rename some module to fix clippy warning
