+++ hospital/src/icu/mod.rs
@@ -1,4 +1,4 @@
-pub mod icu;
+pub mod level;
 
 pub mod proto_icu {
     tonic::include_proto!("icu");
