+++ hospital/src/mix/mod.rs
@@ -1,4 +1,4 @@
-pub mod mix;
+pub mod drees;
 
 pub mod proto_mix {
     tonic::include_proto!("mix");
